Updated leaveData
Updated leaveData, removed placeholder information. It now collects data from the leave.js route and database. It is displaying the Total Quota and Remaining Quota, but not the Used Quota. This might be because it collects a sum of leaves from the total leaves registered and approved in the database, something which does not appear to work at the moment.