Dashboard route moved to new file
The dashboard route was set up in the server.js file. Created a dashboard.js inside the route folder and moved the dashboard route code to it. Added app.use line of code for the dashboard in the server.js file. Also updated the dashboard.ejs from the static leave text to dynamic text as well as update the progress bar for the amount of leaves you have left and made it dynamic.