Changed Leave POST & GET
Changed how the server POST leave requests and GET leave_requests from the database after realizing the employee_id in the leave_request table being tied to the unique id from the employees table was too much of a hassle to work with as it did not match session.id. This forced us to have to work with SQL parameters in order to fetch the user.id.