Updated dashboard.ejs
Added "if" statement to parts of the page to prevent issues loading values of "null". It will check if any of the data fetched and stored inside of activeEntry for the user returns the value "null" If it does, placeholder text will display instead. This is to prevent the issue where a new user who has never clocked in or had breaks before return values with "null". As that was an issue which broke the dashboard page and returned an error 500.