Created Manager Page
Created manager page and corresponding pages and documents to make it so the manager/leave page would collect leave data and display it. The server will first check and see if the  user is assigned as a manager for a department first, if true, then they will be able to view leaves by the employees in that department. If false, they will not gain access to the page.