Updated leaves
Updated the leave management for the manager menu. They can now be approved or rejected. Once either approved or rejected, the leaves can not be changed. User are now able to edit their leave requests as long as their status is set to pending. Once changed to approved or rejected the user is no longer able to make changes to it or delete it.