Updated leaves
Error with user trying to submit a fresh leave request. Ran into an issue with the Row Level Security in Supabase that blocked the insert. This was likely because of our user not carrying a access token that supabase could read, it was only server side. Added a utility file named "supabaseWithSession" which sends the user authorization to supabase with new leave requests to confirm that the user is in the database and is authorized to send leave requests.