Updated departments
Added a button to the view page of departments. It allows the admin to assign a manager to the departments through a modal with a drop down list. The list only display users assigned to that specific department. So if there are no people assigned to the department, then the list will be empty. And by design we have not added a "assign" manager to any new department on the creation page. This is because no one would be in the department to select in a drop down list like this. And we decided on this department specific list to not clutter it with every user in an organization. As that might pose a problem in larger companies. So a department manager will have to be assigned after the creation of a department.